SendInvokeResponse Exception
The ScsServiceApplication class, the method SendInvokeResponse, if an
exception occurs in this method, the block "try catch" reports nothing.
The client application hangs waiting for a response that will never
come. In this case we try to send the same method in reporting the event
an exception (This happens for example when the serialization of an
object of a class not marked as serializable try)