Change ClientId and clients repository
The maximum number of connected clients is given by the maximum capacity
of type long. While this amount is more than enough it is limited. Key
customers change ThreadSafeSortedList of long string still the key to
generating a guid. Thus there is no maximum limit to the number of
customers. The problem does not come from the maximum amount of
concurrent clients but in durable services (eg months) this value could
be achieved if the number of customers is quite large and there are
problems of connection / disconnection frequent.