SSL mutual
Mutual SSL is added. Thus, a client must authenticate against the server
and vice versa. The main advantage of the implemented mechanism is that
the local repository is not used (usually certificates must be installed
using all the trappings of windows). Thus, without directly accessing
the repository PFX files can use. This is useful where many customers
are used and can not have access to the Windows registry